static lib and no warnings, bliss

+++ Callisto/src/Callisto/Imgui/ImguiLayer.cpp
@@ -10,6 +10,7 @@
 #include <GLFW/glfw3.h>
 #include <Glad/glad.h>
 
+
 namespace Callisto
 {
 
@@ -31,7 +32,7 @@ namespace Callisto
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 		// Rendering
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
